requirements_macos.txt
absl-py==2.0.0
annotated-types==0.7.0
anthropic==0.53.0
anyio==4.9.0
asttokens==2.4.1
astunparse==1.6.3
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
cloudpickle==3.0.0
comm==0.2.0
contourpy==1.1.1
ctgan==0.8.0
cycler==0.12.1
decorator==5.1.1
distro==1.9.0
et-xmlfile==1.1.0
eval_type_backport==0.2.2
executing==2.0.1
Faker==19.13.0
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.43.1
fsspec==2023.12.2
gast==0.5.4
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0
google-api-python-client==2.172.0
google-auth==2.23.4
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
google-genai==1.19.0
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.70.0
grpcio==1.73.0
grpcio-status==1.71.0
h11==0.16.0
h5py==3.10.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.4
imbalanced-learn==0.11.0
imblearn==0.0
ipympl==0.9.3
ipython==8.18.1
ipython-genutils==0.2.0
ipywidgets==8.1.1
jedi==0.19.1
Jinja2==3.1.3
jiter==0.10.0
joblib==1.3.2
jupyterlab-widgets==3.0.9
keras==2.14.0
kiwisolver==1.4.5
libclang==16.0.6
llvmlite==0.43.0
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mistralai==1.8.1
ml-dtypes==0.2.0
mpmath==1.3.0
networkx==3.2.1
NICEx==0.2.3
numba==0.60.0
numpy==1.26.1
oauthlib==3.2.2
openai==1.85.0
openpyxl==3.1.4
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.1
parso==0.8.3
pexpect==4.9.0
Pillow==10.1.0
prompt-toolkit==3.0.43
proto-plus==1.26.1
protobuf==5.29.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytorch-tabnet==4.1.0
pytz==2023.3.post1
rdt==1.9.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.4.0
scipy==1.11.3
seaborn==0.13.0
shap==0.46.0
six==1.16.0
slicer==0.0.8
sniffio==1.3.1
stack-data==0.6.3
sympy==1.12
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-macos==2.14.0
termcolor==2.3.0
threadpoolctl==3.2.0
torch==2.1.2
tqdm==4.66.1
traitlets==5.14.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2023.3
uritemplate==4.2.0
urllib3==2.0.7
wcwidth==0.2.12
websockets==15.0.1
Werkzeug==3.0.1
widgetsnbextension==4.0.9
wrapt==1.14.1
xgboost==2.1.1
